refactor: Rearrange functions in `scoring.ts` vs `utils.ts`
