+++ src/lib/utils.ts
@@ -1,5 +1,4 @@
 import { RELATIVE_TIME_CUTOFFS } from "@/lib/constants";
-import type { Player, Session } from "@/types/app.types";
 
 /*
  * Formats a timestamp into human-readable relative time (e.g., "5 minutes ago")
@@ -17,15 +16,6 @@ export function formatTimeAgo(timestamp: string): string {
   }
 
   return rtf.format(delta, "second");
-}
-
-/*
- * Maps a score to its corresponding Tailwind background color class
- * */
-export function scoreToColor(score: number): string {
-  if (score < 0) return "bg-negative";
-  if (score > 0) return "bg-positive";
-  return "bg-neutral";
 }
 
 /*
@@ -55,28 +45,6 @@ export function shuffle<T>(items: T[]): T[] {
   }
 
   return result;
-}
-
-/*
- * Formats a session number into a display name
- * */
-export function getSessionName(session: Session | null): string {
-  if (!session) return "Overall Standings"; // Special case
-
-  return `Session ${session.number} (${session.start_date})`;
-}
-
-/*
- * Combines players with their scores and returns them in descending order
- * */
-export function rankPlayers(
-  players: Player[],
-  scores: Record<string, number>,
-): [Player, number][] {
-  return players
-    .map((player) => ({ player: player, score: scores[player.id] ?? 0 }))
-    .sort((a, b) => b.score - a.score)
-    .map(({ player, score }) => [player, score]);
 }
 
 /*
